Refactor the pkg/util package
