+++ pkg/util/context/context.go
@@ -14,13 +14,15 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package util
+package context
 
 import (
 	"github.com/sirupsen/logrus"
 
 	kubeoneapi "github.com/kubermatic/kubeone/pkg/apis/kubeone"
 	"github.com/kubermatic/kubeone/pkg/ssh"
+	"github.com/kubermatic/kubeone/pkg/util/file"
+	"github.com/kubermatic/kubeone/pkg/util/runner"
 
 	"k8s.io/client-go/rest"
 	dynclient "sigs.k8s.io/controller-runtime/pkg/client"
@@ -32,8 +34,8 @@ type Context struct {
 	Cluster                   *kubeoneapi.KubeOneCluster
 	Logger                    logrus.FieldLogger
 	Connector                 *ssh.Connector
-	Configuration             *Configuration
-	Runner                    *Runner
+	Configuration             *file.Configuration
+	Runner                    *runner.Runner
 	WorkDir                   string
 	JoinCommand               string
 	JoinToken                 string
