+++ pkg/util/runner/runner.go
@@ -14,18 +14,21 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package util
+package runner
 
 import (
+	"bytes"
 	"fmt"
 	"os"
 	"strings"
+	"text/template"
 	"time"
 
 	"github.com/koron-go/prefixw"
 	"github.com/pkg/errors"
 
 	"github.com/kubermatic/kubeone/pkg/ssh"
+	"github.com/kubermatic/kubeone/pkg/util/tee"
 )
 
 // Runner bundles a connection to a host with the verbosity and
@@ -36,6 +39,9 @@ type Runner struct {
 	OS      string
 	Verbose bool
 }
+
+// TemplateVariables is a render context for templates
+type TemplateVariables map[string]interface{}
 
 // Run executes a given command/script, optionally printing its output to
 // stdout/stderr.
@@ -62,8 +68,8 @@ func (r *Runner) Run(cmd string, variables TemplateVariables) (string, string, e
 		return stdout, stderr, err
 	}
 
-	stdout := NewTee(prefixw.New(os.Stdout, r.Prefix))
-	stderr := NewTee(prefixw.New(os.Stderr, r.Prefix))
+	stdout := tee.NewTee(prefixw.New(os.Stdout, r.Prefix))
+	stderr := tee.NewTee(prefixw.New(os.Stderr, r.Prefix))
 
 	// run the command
 	_, err = r.Conn.Stream(cmd, stdout, stderr)
@@ -116,4 +122,19 @@ func (r *Runner) prepareShell(cmd string) string {
 	cmd = fmt.Sprintf("export \"PATH=$PATH:/sbin:/usr/local/bin:/opt/bin\"\n\n%s", cmd)
 
 	return cmd
+}
+
+// MakeShellCommand render text template with given `variables` render-context
+func MakeShellCommand(cmd string, variables TemplateVariables) (string, error) {
+	tpl, err := template.New("base").Parse(cmd)
+	if err != nil {
+		return "", errors.Wrap(err, "failed to parse shell script")
+	}
+
+	buf := bytes.Buffer{}
+	if err := tpl.Execute(&buf, variables); err != nil {
+		return "", errors.Wrap(err, "failed to render shell script")
+	}
+
+	return buf.String(), nil
 }
