--- pkg/util/util.go
@@ -1,42 +1,0 @@
-/*
-Copyright 2019 The KubeOne Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package util
-
-import (
-	"bytes"
-	"text/template"
-
-	"github.com/pkg/errors"
-)
-
-// TemplateVariables is a render context for templates
-type TemplateVariables map[string]interface{}
-
-// MakeShellCommand render text template with given `variables` render-context
-func MakeShellCommand(cmd string, variables TemplateVariables) (string, error) {
-	tpl, err := template.New("base").Parse(cmd)
-	if err != nil {
-		return "", errors.Wrap(err, "failed to parse shell script")
-	}
-
-	buf := bytes.Buffer{}
-	if err := tpl.Execute(&buf, variables); err != nil {
-		return "", errors.Wrap(err, "failed to render shell script")
-	}
-
-	return buf.String(), nil
-}
